refactor(formula): function simply
