+++ packages/engine-formula/src/engine/reference-object/base-reference-object.ts
@@ -21,9 +21,12 @@ import { FormulaAstLRU } from '../../basics/cache-lru';
 import type { INumfmtItemMap, IRuntimeUnitDataType, IUnitData, IUnitSheetNameMap } from '../../basics/common';
 import { ERROR_TYPE_SET, ErrorType } from '../../basics/error-type';
 import { ObjectClassType } from '../../basics/object-class-type';
-import { ArrayValueObject, createBooleanValueObjectByRawValue, createNumberValueObjectByRawValue, createStringValueObjectByRawValue, ValueObjectFactory } from '../value-object/array-value-object';
+import { ArrayValueObject, ValueObjectFactory } from '../value-object/array-value-object';
 import { type BaseValueObject, ErrorValueObject, type IArrayValueObject } from '../value-object/base-value-object';
 import {
+    createBooleanValueObjectByRawValue,
+    createNumberValueObjectByRawValue,
+    createStringValueObjectByRawValue,
     NullValueObject,
     NumberValueObject,
 } from '../value-object/primitive-object';
